refactor(ai-chat): move the editor storage path context out of sessions
